% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapCI.R
\name{BootstrapCI}
\alias{BootstrapCI}
\alias{bootstrapCI,}
\alias{bootsCI,}
\alias{BootsCI}
\title{Extract Four Sets of Confidence Intervals (CI) from the Bootstrap Results}
\usage{
BootstrapCI(t.bootstrap, N.Bn, t.0 = NULL, trace = FALSE,
  conf.level = 0.95, print.it = TRUE)
}
\arguments{
\item{t.bootstrap}{A vector of bootstrapped statistics of dimention \eqn{B}.}

\item{N.Bn}{The incidence matrix of dim \eqn{latex}{B x n} associated with the bootstrap
resampling procedure, where element \eqn{latex}{B_{bi}} is the frequency that the \eqn{i}-th observation
shows up in the \eqn{b}-th bootstrap sample.}

\item{t.0}{The statistics computed on basis of the original sample.}

\item{trace}{A logical value. If \code{TRUE}, information on progress is shown. The default is \code{FALSE}.}

\item{conf.level}{A numerical value with the confidence level for the construction of the
confidence intervals. The default value is 0.95.}

\item{print.it}{A logical value. If \code{TRUE}, a summary of results is printed. Defaulted as \code{TRUE}.}
}
\value{
The following results are returned:
\describe{
\item{result}{A vector of results including the sample estimate \code{t.0}, its bootstrap SE \code{SE.bootstrap},
      the lower/upper bound of the bootstrap CI (\code{LB1.bootstrap}, \code{UB1.bootstrap}), the bagging
      estimate \code{t.bagged}, the IJ-based SE \code{SE.IJ}, the lower/upper bound of the IJ-based CI
      (\code{LB1.IJ}, \code{UB1.IJ}), the bias estimate \code{bias} for variance, the bias-corrected IJ variance
      estimate \code{Vc.IJ}, the bias-corrected IJ SE \code{SEc.IJ}, the lower/upper bound of the bias-corrected
      IJ confidence interval (\code{LB2.IJ}, \code{UB2.IJ}).}
\item{t.bootstrap}{A copy of the \eqn{B}-dimensional vector of bootstrapped statistics;}
\item{N.Bn}{A copy of the bootstrap incidence matrix;}
\item{V.IJ}{The IJ estimate of the variance;}
\item{bias}{The (upward) bias estimate for \code{V.IJ};}
\item{Vc.IJ}{The bias-corrected IJ variance estimate;}
\item{conf.level}{Confidence level used;}
\item{B}{The number of bootstrap samples used.}
}
}
\description{
Extract Four Sets of Confidence Intervals (CI) from the Bootstrap Results
}
\details{
Four sets of confidence intervals are extracted: the bootstrap confidence interval and the percentile CI for
      the sample estimate of the optimal cutpoint; the infinitesimal jackknife (IJ) confidence interval and its
      bias-corrected version for the bagging estimator of the optimal cutpoint. The bagging estimator is more
      precise with smaller variation. The bias correction is helpful when \eqn{B} is small.

      This function is used inside of Function \code{LogitOptimalCutpointsCI} in this package, but it is a generaic function that
      can be used for other general purposes.
}
\examples{
      data(fertility)
      form0 <- diagnosis ~  factor(season) + age + accident + fever + alcohol + sitting
      out.boots <- BootstrapLogitBestcut(formula=form0, dat=fertility, B=100,
                 trace=FALSE, conf.level=0.95, CV.method="none",  method.OptimalCutpoints = "Youden")
      out <- BootstrapCI(out.boots$c.bootstrap, out.boots$N.Bn, t.0=out.boots$c.0, trace=FALSE,
                 conf.level=0.95, print.it=TRUE)
      out
      plot.BootstrapCI(out)

}
\references{
\itemize{
\item Zhang, Z., Su, X., et al. (2019+). Bootstrap Confidence Intervals for Optimal Cutoff Point
in Logistic Regression. Submitted.
}
}
\seealso{
\code{\link{BootstrapLogitBestcut}}, \code{\link{OptimalThreshold}}, \code{\link{plot.BootstrapCI}}
}
