% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapLogitBestcut.R
\name{BootstrapLogitBestcut}
\alias{BootstrapLogitBestcut}
\alias{BootstrapLogitOptimalCutpoint}
\title{Bootstrapping the best cutoff points with logistic regression}
\usage{
BootstrapLogitBestcut(formula, dat, B = 2000, trace = FALSE,
  conf.level = 0.95, cntr.glm = glm.control(epsilon = 1e-08, maxit =
  100, trace = FALSE), seed = 123, stratified.sampling = FALSE,
  CV.method = "none", V.fold = NULL, cntr0 = control.cutpoints(),
  method.OptimalCutpoints = "Youden")
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}, with the response on the left of
a \code{~} operator, and the terms on the right.}

\item{dat}{A data.frame in which to interpret the variables named in the \code{formula} argument.}

\item{B}{The number of bootstrap samples used. Suggest setting \eqn{B} large such as 1000 02 2000.}

\item{trace}{A logical value. If \code{TRUE}, information on progress is shown. The default is \code{FALSE}.}

\item{conf.level}{A numerical value with the confidence level for the construction of the
confidence intervals. The default value is 0.95.}

\item{cntr.glm}{The control setting for fitting logistic regression model with \code{glm}.
See \code{\link[stats]{glm.control}} for details.}

\item{seed}{The ranomd seed to reproduce results. Default is 123.}

\item{stratified.sampling}{A logical value indicating whether stratified sampling should be used in bootstrap
resampling and V-fold CV with stratification on the response. It is highly recommended with unbalanced
classification problems.}

\item{CV.method}{Values in \code{"none"}, \code{"OOB"}, \code{"LOO"}. Default is \code{"none"} with
no cross validation in computing the predicted probabilities. Otherwise, either \code{"OOB"} out-of-bag
samples or \code{"LOO"} leave-one-out jackknife method is used.}

\item{V.fold}{The number of folds in V-fold CV only when \code{CV.method="OOB"}. In this case, V-fold CV
is used to compute the predicted probabilities with the orinal sample data. The choice of \code{V.fold=3} is
suggested, as 1/3 = .33 roughly corresponds to .37 in OOB. Even so, the bootstrap/percentile CIs are still
doubtful.}

\item{cntr0}{The control setting for \code{\link[OptimalCutpoints]{optimal.cutpoints}}, modified as
\code{OptimalThreshold}. See \code{\link[OptimalCutpoints]{control.cutpoints}} for details.}

\item{method.OptimalCutpoints}{A character vector selecting the method(s) to be used in \code{OptimalThreshold};
See \code{\link[OptimalCutpoints]{optimal.cutpoints}} for details. Choices are "CB" (cost-benefit method);
"MCT" (minimizes Misclassification Cost Term); "MinValueSp" (a minimum value set for Specificity);
"MinValueSe" (a minimum value set for Sensitivity); "ValueSe" (a value set for Sensitivity);
"MinValueSpSe" (a minimum value set for Specificity and Sensitivity); "MaxSp" (maximizes Specificity);
"MaxSe" (maximizes Sensitivity);"MaxSpSe" (maximizes Sensitivity and Specificity simultaneously);
"MaxProdSpSe" (maximizes the product of Sensitivity and Specificity or Accuracy Area);
"ROC01" (minimizes distance between ROC plot and point (0,1)); "SpEqualSe" (Sensitivity = Specificity);
"Youden" (Youden Index); "MaxEfficiency" (maximizes Efficiency or Accuracy, similar to minimize Error
Rate); "Minimax" (minimizes the most frequent error); "MaxDOR" (maximizes Diagnostic Odds Ratio);
"MaxKappa" (maximizes Kappa Index); "MinValueNPV" (a minimum value set for Negative Predictive Value);
"MinValuePPV" (a minimum value set for Positive Predictive Value); "ValueNPV" (a value set for Negative
Predictive Value);"ValuePPV" (a value set for Positive Predictive Value);"MinValueNPVPPV" (a minimum
value set for Predictive Values); "PROC01" (minimizes distance between PROC plot and point (0,1));
"NPVEqualPPV" (Negative Predictive Value = Positive Predictive Value); "MaxNPVPPV" (maximizes Positive
Predictive Value and Negative Predictive Value simultaneously); "MaxSumNPVPPV" (maximizes the sum of
the Predictive Values); "MaxProdNPVPPV" (maximizes the product of Predictive Values); "ValueDLR.Negative"
(a value set for Negative Diagnostic Likelihood Ratio); "ValueDLR.Positive" (a value set for Positive
Diagnostic Likelihood Ratio); "MinPvalue" (minimizes p-value associated with the statistical Chi-squared
test which measures the association between the marker and the binary result obtained on using the
cutpoint); "ObservedPrev" (The closest value to observed prevalence); "MeanPrev" (The closest value
to the mean of the diagnostic test values); or "PrevalenceMatching" (The value for which predicted
prevalence is practically equal to observed prevalence).

#' @details
This immediate funciton runs a bootstrapping procedure for obtaining a bootstrap sample of the optimal cutpoint
with logistic regression. Its output will be used as input for \code{\link{BootstrapCI}}.}
}
\value{
Four objects are retuned.
\describe{
\item{result.0}{The object from \code{OptimalThreshold}, same as \code{\link[OptimalCutpoints]{optimal.cutpoints}},
for the original sample data.}
\item{c.0}{The optimal cutoff point based on the original sample data;}
\item{c.bootstrap}{The bagging estimator for the optimal cutpoint;}
\item{N.Bn}{The incidence matrix of dim \eqn{latex}{B x n} associated with the bootstrap
resampling procedure, where element \eqn{latex}{B_{bi}} is the frequency that the \eqn{i}-th observation
shows up in the \eqn{b}-th bootstrap sample.}
}
}
\description{
Bootstrapping the best cutoff points with logistic regression
}
\examples{
      data(fertility)
      form0 <- diagnosis ~  factor(season) + age + accident + fever + alcohol + sitting
      out.boots <- BootstrapLogitBestcut(formula=form0, dat=fertility, B=100,
                 trace=FALSE, conf.level=0.95, CV.method="none", V.fold=3,
                 cntr0=control.cutpoints(), method.OptimalCutpoints = "Youden")
      names(out.boots)

}
\references{
\itemize{
\item Zhang, Z., Su, X., et al. (2019+). Bootstrap Confidence Intervals for Optimal Cutoff Point
in Logistic Regression. Submitted.
}
}
\seealso{
\code{\link[OptimalCutpoints]{optimal.cutpoints}}, \code{\link{OptimalThreshold}}
}
