% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalThreshold.R
\name{OptimalThreshold}
\alias{OptimalThreshold}
\alias{optimal.threshold}
\title{A function modified from Function \code{optimal.cutpoints} in the R Package \bold{OptimalCutpoints}}
\usage{
OptimalThreshold(X, status = NULL, tag.healthy, methods, data,
  direction = c("<", ">"), categorical.cov = NULL, pop.prev = NULL,
  control = control.cutpoints(), ci.fit = FALSE, conf.level = 0.95,
  trace = FALSE, ...)
}
\arguments{
\item{X}{Either a character string with the name of the diagnostic test variable (then method
'optimal.cutpoints.default' is called), or a formula (then method 'optimal.cutpoints.formula'
is called). When 'X' is a formula, it must be an object of class "formula". Right side of \code{~}
must contain the name of the variable that distinguishes healthy from diseased individuals, and
left side of \code{~} must contain the name of the diagnostic test variable.}

\item{status}{A character string with the name of the variable that distinguishes healthy from
diseased individuals. Only applies for the method 'optimal.cutpoints.default').}

\item{tag.healthy}{the value codifying healthy individuals in the status variable.}

\item{methods}{A character vector selecting the method(s) to be used: "CB" (cost-benefit method);
"MCT" (minimizes Misclassification Cost Term); "MinValueSp" (a minimum value set for Specificity);
"MinValueSe" (a minimum value set for Sensitivity); "ValueSe" (a value set for Sensitivity);
"MinValueSpSe" (a minimum value set for Specificity and Sensitivity); "MaxSp" (maximizes Specificity);
"MaxSe" (maximizes Sensitivity);"MaxSpSe" (maximizes Sensitivity and Specificity simultaneously);
"MaxProdSpSe" (maximizes the product of Sensitivity and Specificity or Accuracy Area);
"ROC01" (minimizes distance between ROC plot and point (0,1)); "SpEqualSe" (Sensitivity = Specificity);
"Youden" (Youden Index); "MaxEfficiency" (maximizes Efficiency or Accuracy, similar to minimize Error
Rate); "Minimax" (minimizes the most frequent error); "MaxDOR" (maximizes Diagnostic Odds Ratio);
"MaxKappa" (maximizes Kappa Index); "MinValueNPV" (a minimum value set for Negative Predictive Value);
"MinValuePPV" (a minimum value set for Positive Predictive Value); "ValueNPV" (a value set for Negative
Predictive Value);"ValuePPV" (a value set for Positive Predictive Value);"MinValueNPVPPV" (a minimum
value set for Predictive Values); "PROC01" (minimizes distance between PROC plot and point (0,1));
"NPVEqualPPV" (Negative Predictive Value = Positive Predictive Value); "MaxNPVPPV" (maximizes Positive
Predictive Value and Negative Predictive Value simultaneously); "MaxSumNPVPPV" (maximizes the sum of
the Predictive Values); "MaxProdNPVPPV" (maximizes the product of Predictive Values); "ValueDLR.Negative"
(a value set for Negative Diagnostic Likelihood Ratio); "ValueDLR.Positive" (a value set for Positive
Diagnostic Likelihood Ratio); "MinPvalue" (minimizes p-value associated with the statistical Chi-squared
test which measures the association between the marker and the binary result obtained on using the
cutpoint); "ObservedPrev" (The closest value to observed prevalence); "MeanPrev" (The closest value
to the mean of the diagnostic test values); or "PrevalenceMatching" (The value for which predicted
prevalence is practically equal to observed prevalence).}

\item{data}{A data frame containing all needed variables.}

\item{direction}{A Character string specifying the direction to compute the ROC curve. By default
individuals with a test value lower than the cutoff are classified as healthy (negative test),
whereas patients with a test value greater than (or equal to) the cutoff are classified as
diseased (positive test). If this is not the case, however, and the high values are related to
health, this argument should be established at ">".}

\item{categorical.cov}{A character string with the name of the categorical covariate according
to which optimal cutpoints are to be calculated. The default is NULL (no categorical covariate).}

\item{pop.prev}{The value of the disease's prevalence. The default is \code{NULL} (prevalence is
estimated on the basis of sample prevalence). It can be a vector indicating the prevalence values
for each categorical covariate level.}

\item{control}{Output of the \code{\link[OptimalCutpoints]{control.cutpoints}}  function.}

\item{ci.fit}{A logical value. If \code{TRUE}, inference is performed on the accuracy measures
at the optimal cutpoint. The default is \code{FALSE}.}

\item{conf.level}{A numerical value with the confidence level for the construction of the
confidence intervals. The default value is 0.95.}

\item{trace}{A logical value. If \code{TRUE}, information on progress is shown. The default is \code{FALSE}.}

\item{\code{...}}{Further arguments passed to or from other methods. None are used in this method.}
}
\value{
Returns an object of class "optimal.cutpoints" with the following components: \code{methods},
\code{levels.cat}, \code{call}, \code{dat}, etc. See \code{\link[OptimalCutpoints]{optimal.cutpoints}}
for details.
}
\description{
A function modified from Function \code{optimal.cutpoints} in the R Package \bold{OptimalCutpoints}
}
\details{
See \code{\link[OptimalCutpoints]{optimal.cutpoints}}.
}
\examples{
   beta0 <- c(-2, 2, -2, 2, -2, -2)
   dat <- rdat(n=500, beta=beta0)
   fit <- glm(y ~., data=dat, family=binomial(link="logit"),
              control=glm.control(epsilon=1e-8, maxit=100, trace=FALSE))
   summary(fit)
   dat.test <- rdat(n=500, beta=-beta0)
   pred <- predict(fit, newdata=dat.test, type="response")
   dat.tmp <- data.frame(pred=pred, y=dat.test$y)
   result <- OptimalCut(X="pred", status="y", data=dat.tmp, tag.healthy=0, methods = "Youden")
   c0 <- result$output$Global$optimal.cutoff$cutoff;
   c0  # THE BEST CUTOFF POINT
}
\references{
\itemize{
\item Lopez-Raton, M., Rodriguez-Alvarez, M. X., Cadarso-Suarez, C., and Gude-Sampedro, F. (2014).
OptimalCutpoints: An R package for selecting optimal cutpoints in diagnostic tests. \emph{Journal of
Statistical Software}, \bold{61}(8), 1. URL \url{http://www.jstatsoft.org/v61/i08/}.
\item Zhang, Z., Su, X., et al. (2019+). Bootstrap Confidence Intervals for Optimal Cutoff Point
in Logistic Regression. Submitted.
}
}
\seealso{
\code{\link[OptimalCutpoints]{optimal.cutpoints}}
}
