% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-BootstrapCI.R
\name{plot.BootstrapCI}
\alias{plot.BootstrapCI}
\alias{plot.bootstrapCI}
\title{The Generic \code{plot} Function for Object of \code{BootstrapCI} Class}
\usage{
\method{plot}{BootstrapCI}(cutpoint,
  main = "Histogram of Bootstrapped Optimal Cutpoints", cex.text = 0.8,
  a0 = 5, b0 = a0 - 1 + 0.2, gap = 0.2)
}
\arguments{
\item{cutpoint}{an object of \code{BootstrapCI} class, returned from Function \code{BootstrapCI}.}

\item{main}{The title of the plot.}

\item{cex.text}{The font size for labeling the four sets of bootstrap CIs.}

\item{a0, b0, gap}{Figure parameters for setting some margins in the plot to make it look nicers.
No need to adjust in general.}
}
\value{
Plots of the results from \code{BootstrapCI}.
}
\description{
The Generic \code{plot} Function for Object of \code{BootstrapCI} Class
}
\details{
The (generic) plot method for an \code{BootstrapCI} object. It plots the histogram of the bootstrapped statistics,
together with four sets of bootstrap confidence intervals.
}
\examples{
      data(fertility)
      form0 <- diagnosis ~  factor(season) + age + accident + fever + alcohol + sitting
      out.boots <- BootstrapLogitBestcut(formula=form0, dat=fertility, B=100,
                 trace=FALSE, conf.level=0.95, CV.method="none",  method.OptimalCutpoints = "Youden")
      out <- BootstrapCI(out.boots$c.bootstrap, out.boots$N.Bn, t.0=out.boots$c.0, trace=FALSE,
                 conf.level=0.95, print.it=TRUE)
      out
      plot(out)

}
\references{
\itemize{
\item Zhang, Z., Su, X., et al. (2019+). Bootstrap Confidence Intervals for Optimal Cutoff Point
in Logistic Regression. Submitted.
}
}
\seealso{
\code{\link{BootstrapLogitBestcut}}, \code{\link{BootstrapCI}}
}
