% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-BootstrapCI.R
\name{print.BootstrapCI}
\alias{print.BootstrapCI}
\alias{print.bootstrapCI}
\title{The Generic \code{print} Function for Object of \code{BootstrapCI} Class}
\usage{
\method{print}{BootstrapCI}(cutpoint)
}
\arguments{
\item{cutpoint}{an object of \code{BootstrapCI} class, returned from Function \code{BootstrapCI}.}
}
\description{
The Generic \code{print} Function for Object of \code{BootstrapCI} Class
}
\examples{
      data(fertility)
      form0 <- diagnosis ~  factor(season) + age + accident + fever + alcohol + sitting
      out.boots <- BootstrapLogitBestcut(formula=form0, dat=fertility, B=100,
                 trace=FALSE, conf.level=0.95, CV.method="none",  method.OptimalCutpoints = "Youden")
      out <- BootstrapCI(out.boots$c.bootstrap, out.boots$N.Bn, t.0=out.boots$c.0, trace=FALSE,
                 conf.level=0.95, print.it=TRUE)
      print(out)

}
\references{
\itemize{
\item Zhang, Z., Su, X., et al. (2019+). Bootstrap Confidence Intervals for Optimal Cutoff Point
in Logistic Regression. Submitted.
}
}
\seealso{
\code{\link{BootstrapLogitBestcut}}, \code{\link{BootstrapCI}}
}
