% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdat.R
\name{rdat}
\alias{rdat}
\alias{rdata}
\title{Generate some data from a logistic regression model}
\usage{
rdat(n, beta, x.dist = "uniform")
}
\arguments{
\item{n}{The sample size.}

\item{beta}{The vector of regression coefficients. The first component is intercept.
The number of predictors is automatically determined by the dimension of beta}

\item{x.dist}{Predictors are independently generated from either \code{"uniform"} or \code{"normal"}
distributions. Correlated predictors could have been easily used, but we didn't bother to do so.}
}
\value{
A data frame \code{dat} where the first column is the binary response \code{y}, followed by
a number of predictors \code{x1}, \code{x2}, \eqn{\ldots}.
}
\description{
Generate some data from a logistic regression model
}
\details{
The function simualtes data from a true logistic regression model. The empirical prevalence rate for
\code{y=1} is printed out.
}
\examples{
      beta0 <- c(-2, 2, -2, 2, -2, -2)
      dat <- rdat(n=1000, beta=beta0)
      dim(dat); head(dat)

}
\references{
\itemize{
\item Zhang, Z., Su, X., et al. (2019+). Bootstrap Confidence Intervals for Optimal Cutoff Point
in Logistic Regression. Submitted.
}
}
\seealso{
\code{\link{rnorm}}, \code{\link{runif}}
}
